moved config into a module. #495
