+++ src/js/global-dependencies.js
@@ -35,8 +35,8 @@ import * as ThemeController from "./theme-controller";
 import * as Caret from "./caret";
 import * as CustomTextPopup from "./custom-text-popup";
 import * as ManualRestart from "./manual-restart-tracker";
-import Config from "./config";
-import * as ConfigSet from "./config-set";
+import Config, * as UpdateConfig from "./config";
+// import * as ConfigSet from "./config-set";
 import * as Focus from "./focus";
 import * as AccountIcon from "./account-icon";
 import * as PractiseMissed from "./practise-missed";
@@ -51,3 +51,8 @@ import * as TestLogic from "./test-logic";
 import * as Funbox from "./funbox";
 import * as PaceCaret from "./pace-caret";
 import * as QuoteSearchPopup from "./quote-search-popup";
+import * as TagController from "./tag-controller";
+import * as TestTimer from "./test-timer";
+import * as LanguagePicker from "./language-picker";
+import * as UI from "./ui";
+import * as PbCrown from "./pb-crown";
