+++ src/js/ui.js
@@ -1,0 +1,5 @@
+export let pageTransition = false;
+
+export function setPageTransition(val) {
+  pageTransition = val;
+}
